requirements.txt
Cython==0.29.2
hidapi==0.7.99.post21
pyusb==1.0.2
